pyproject.toml
[project]
name = "03-batch-file-convertor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pillow>=11.3.0",
    "pymupdf>=1.26.5",
    "img2pdf>=0.6.3",
    "streamlit>=1.50.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--cov=services",
    "--cov=core",
    "--cov-report=term-missing",
    "--cov-fail-under=80",
]
